Allow a wider range of dateutil versions
By requiring only an exact version of python-dateutil, many projects
were unable to resolve a dateutil version that works for all
dependencies. Now that there is a spec in place which tests the
only usage of this package, I was able to validate that upgrading
dateutil for any 2.X greater than 2.4.2 works.